requirements.txt
aiohttp==3.8.4
aiohttp-cors==0.7.0
aiokafka==0.8.1
aiosignal==1.3.1
async-timeout==4.0.2
asyncpg==0.27.0
attrs==23.1.0
Automat==22.10.0
beautifulsoup4==4.12.2
boto3==1.26.115
botocore==1.29.153
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
colorlog==6.7.0
constantly==15.1.0
croniter==1.3.15
cryptography==41.0.1
cssselect==1.2.0
elastic-transport==8.4.0
elasticsearch==8.8.0
faust-streaming==0.10.0
filelock==3.12.2
frozenlist==1.3.3
furl==2.1.3
greenlet==2.0.2
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
intervaltree==3.1.0
itemadapter==0.8.0
itemloaders==1.1.0
jmespath==1.0.1
kafka-python==2.0.2
lxml==4.9.2
mode-streaming==0.3.5
multidict==6.0.4
mypy-extensions==1.0.0
opentracing==2.4.0
orderedmultidict==1.0.1
packaging==23.1
parsel==1.8.1
Protego==0.2.1
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.2
PyDispatcher==2.0.7
pyOpenSSL==23.2.0
python-dateutil==2.8.2
python-json-logger==2.0.4
queuelib==1.6.2
requests==2.31.0
requests-file==1.5.1
s3transfer==0.6.1
Scrapy==2.8.0
selenium==4.10.0
service-identity==23.1.0
six==1.16.0
sortedcontainers==2.4.0
soupsieve==2.4.1
SQLAlchemy==2.0.16
terminaltables==3.1.10
tldextract==3.4.4
Twisted==22.10.0
typing_extensions==4.6.3
undetected-chromedriver==3.5.0
urllib3==1.26.16
venusian==3.0.0
w3lib==2.1.1
yarl==1.9.2
zope.interface==6.0
